refactor: share the title bar passthrough regions across windows
